+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpInteropTest.kt
@@ -1,7 +1,5 @@
 package at.asitplus.wallet.lib.openid
 
-import at.asitplus.dif.ClaimFormat
-import at.asitplus.dif.PresentationSubmission
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.RequestParametersFrom
@@ -47,6 +45,8 @@ import io.kotest.matchers.string.shouldStartWith
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.http.*
 import kotlinx.coroutines.runBlocking
+import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonPrimitive
 
 /**
@@ -133,7 +133,7 @@ val OpenId4VpInteropTest by matrixSuite {
                                 DCQLClaimsPathPointer(CLAIM_GIVEN_NAME)
                             )
                         )
-                    ).toPresentationExchangeRequest(),
+                    ).toDCQLRequest(),
                     responseMode = OpenIdConstants.ResponseMode.DirectPost,
                     responseUrl = "https://verifier.example.com/response/$responseNonce",
                 ),
@@ -167,7 +167,7 @@ val OpenId4VpInteropTest by matrixSuite {
             jarPayload.audience shouldBe "https://self-issued.me/v2"
             jarPayload.clientId shouldBe it.verifierClientId
             jarPayload.clientIdWithoutPrefix shouldBe it.verifierClientId
-            jarPayload.presentationDefinition.shouldNotBeNull()
+            jarPayload.dcqlQuery.shouldNotBeNull()
             jarPayload.nonce.shouldNotBeNull()
             jarPayload.state.shouldNotBeNull()
             jarPayload.responseType shouldBe "vp_token"
@@ -185,7 +185,11 @@ val OpenId4VpInteropTest by matrixSuite {
                 .shouldBeInstanceOf<AuthenticationResponseResult.Post>()
 
             response.params.entries.firstOrNull { it.key == "vp_token" }.shouldNotBeNull().value.let { vpToken ->
-                val sdJwt = SdJwtSigned.parseCatching(vpToken).getOrThrow()
+                val presentation = joseCompliantSerializer.decodeFromString<JsonObject>(vpToken)
+                    .values.shouldBeSingleton().first()
+                    .jsonArray.shouldBeSingleton().first()
+                    .jsonPrimitive.content
+                val sdJwt = SdJwtSigned.parseCatching(presentation).getOrThrow()
                 sdJwt.keyBindingJws.shouldNotBeNull().apply {
                     jws.jwsHeader.apply {
                         algorithm shouldBe JwsAlgorithm.Signature.ES256
@@ -218,19 +222,13 @@ val OpenId4VpInteropTest by matrixSuite {
                 }
             }
             response.params.entries.firstOrNull { it.key == "state" }.shouldNotBeNull()
-            response.params.entries.first { it.key == "presentation_submission" }.value.let { presentationSubmission ->
-                val presSub = joseCompliantSerializer.decodeFromString<PresentationSubmission>(presentationSubmission)
-                presSub.definitionId.shouldNotBeNull()
-                presSub.descriptorMap.shouldNotBeNull().first().apply {
-                    path shouldBe "$"
-                    format shouldBe ClaimFormat.SD_JWT
-                }
-            }
+            response.params.entries.firstOrNull { it.key == "presentation_submission" }.shouldBeNull()
 
             it.verifierOid4vp.validateAuthnResponse(response.params.formUrlEncode()).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.shouldBeSingleton().first().getOrThrow()
+                .shouldBeInstanceOf<VpTokenValidationResultDCQL>()
+                .credentialQueryResponseValidations.values.shouldBeSingleton().first()
+                .shouldBeSingleton().first().getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
         }