+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/KeyBindingTests.kt
@@ -1,7 +1,6 @@
 package at.asitplus.wallet.lib.rqes
 
 import at.asitplus.csc.contentEquals
-import at.asitplus.iso.sha256
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.QCertCreationAcceptance
@@ -21,7 +20,6 @@ import at.asitplus.wallet.lib.agent.ValidatorSdJwt
 import at.asitplus.wallet.lib.agent.Verifier
 import at.asitplus.wallet.lib.agent.VerifierAgent
 import at.asitplus.wallet.lib.agent.toStoreCredentialInput
-import at.asitplus.wallet.lib.NonceService
 import at.asitplus.wallet.lib.data.AttributeIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
 import at.asitplus.wallet.lib.data.SdJwtConstants
@@ -34,21 +32,18 @@ import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
 import at.asitplus.wallet.lib.openid.ClientIdScheme
 import at.asitplus.wallet.lib.openid.OpenId4VpHolder
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier
-import at.asitplus.wallet.lib.openid.VpTokenValidationResultPresentationExchange
+import at.asitplus.wallet.lib.openid.VpTokenValidationResultDCQL
 import at.asitplus.wallet.lib.rqes.helper.DummyCredentialDataProvider
 import at.asitplus.wallet.lib.utils.DefaultMapStore
 import com.benasher44.uuid.bytes
 import com.benasher44.uuid.uuid4
 import io.kotest.matchers.collections.shouldBeSingleton
-import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.http.*
-import io.ktor.utils.io.charsets.*
-import io.ktor.utils.io.core.*
 import kotlinx.coroutines.runBlocking
 
 private fun malignTransactionData(): List<TransactionDataBase64Url> = listOf(
@@ -87,110 +82,6 @@ val KeyBindingTests by matrixSuite {
 
                 val walletUrl = "https://example.com/wallet/${uuid4()}"
                 val clientId = "https://example.com/rp/${uuid4()}"
-                val cibaWalletTransactionData = """
-                eyJ0eXBlIjoicWNlcnRfY3JlYXRpb25fYWNjZXB0YW5jZSIsImNyZWRlbnRpYWxfaWRzIjpbIjYwNzUxMGE5LWM5NTctNDA5NS05MDZkLWY5
-                OWZkMDA2YzRhZSJdLCJRQ190ZXJtc19jb25kaXRpb25zX3VyaSI6Imh0dHBzOi8vd3d3LmQtdHJ1c3QubmV0L2RlL2FnYiIsIlFDX2hhc2gi
-                OiI3UXptNUVqdXpYS1NIRmxjME9IOVBQOXFVYUgtVkJsMmFHTmJ3WWoxb09BIiwiUUNfaGFzaEFsZ29yaXRobU9JRCI6IjIuMTYuODQwLjEu
-                MTAxLjMuNC4yLjEiLCJ0cmFuc2FjdGlvbl9kYXRhX2hhc2hlc19hbGciOlsic2hhLTI1NiJdfQ
-            """.trimIndent().replace("\n", "")
-
-                val cibaWalletTestVector = """
-                    {
-                        "response_type": "vp_token",
-                        "client_id": "redirect_uri:$clientId",
-                        "scope": "",
-                        "state": "iTGlKl-AJxmncWPbXHp2xy58bNy18wqZ4TR9EzhBl2R4ulxeTEO0VyWYR2qMDpCDV5JWeOxecTqcEJ61bFKrUg",
-                        "nonce": "f90d0982-52f4-4a1c-8525-bdf1d33c232b",
-                        "client_metadata": {
-                            "jwks_uri": "https://cibawallet.local-ip.medicmobile.org/wallet/jarm/iTGlKl-AJxmncWPbXHp2xy58bNy18wqZ4TR9EzhBl2R4ulxeTEO0VyWYR2qMDpCDV5JWeOxecTqcEJ61bFKrUg/jwks.json",
-                            "id_token_signed_response_alg": "RS256",
-                            "authorization_encrypted_response_alg": "ECDH-ES",
-                            "authorization_encrypted_response_enc": "A128CBC-HS256",
-                            "id_token_encrypted_response_alg": "RSA-OAEP-256",
-                            "id_token_encrypted_response_enc": "A128CBC-HS256",
-                            "subject_syntax_types_supported": [
-                                "urn:ietf:params:oauth:jwk-thumbprint"
-                            ],
-                            "vp_formats": {
-                                "dc+sd-jwt": {
-                                    "sd-jwt_alg_values": [
-                                        "ES256"
-                                    ],
-                                    "kb-jwt_alg_values": [
-                                        "ES256"
-                                    ]
-                                },
-                                "dc+sd-jwt": {
-                                    "sd-jwt_alg_values": [
-                                        "ES256"
-                                    ],
-                                    "kb-jwt_alg_values": [
-                                        "ES256"
-                                    ]
-                                },
-                                "mso_mdoc": {
-                                    "alg": [
-                                        "ES256"
-                                    ]
-                                }
-                            }
-                        },
-                        "presentation_definition": {
-                            "id": "4c7038cf-bd1e-47c0-8f70-eaf9d62c6fae",
-                            "name": "Cibazmaj",
-                            "purpose": "where su pare",
-                            "input_descriptors": [
-                                {
-                                    "id": "607510a9-c957-4095-906d-f99fd006c4ae",
-                                    "name": "niko kao",
-                                    "purpose": "hajduk iz splita",
-                                    "format": {
-                                        "dc+sd-jwt": {
-                                            "sd-jwt_alg_values": [
-                                                "ES256"
-                                            ],
-                                            "kb-jwt_alg_values": [
-                                                "ES256"
-                                            ]
-                                        }
-                                    },
-                                    "constraints": {
-                                        "fields": [
-                                            {
-                                                "path": [
-                                                    "${'$'}.family_name"
-                                                ]
-                                            },
-                                            {
-                                                "path": [
-                                                    "${'$'}.given_name"
-                                                ]
-                                            },
-                                            {
-                                                "path": [
-                                                    "${'$'}.vct"
-                                                ],
-                                                "filter": {
-                                                    "type": "string",
-                                                    "enum": [
-                                                        "urn:eudi:pid:1"
-                                                    ]
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            ]
-                        },
-                        "response_mode": "direct_post",
-                        "response_uri": "https://cibawallet.local-ip.medicmobile.org/wallet/direct_post/iTGlKl-AJxmncWPbXHp2xy58bNy18wqZ4TR9EzhBl2R4ulxeTEO0VyWYR2qMDpCDV5JWeOxecTqcEJ61bFKrUg",
-                        "aud": "https://self-issued.me/v2",
-                        "iat": 1744198186,
-                        "transaction_data": [
-                            "eyJ0eXBlIjoicWNlcnRfY3JlYXRpb25fYWNjZXB0YW5jZSIsImNyZWRlbnRpYWxfaWRzIjpbIjYwNzUxMGE5LWM5NTctNDA5NS05MDZkLWY5OWZkMDA2YzRhZSJdLCJRQ190ZXJtc19jb25kaXRpb25zX3VyaSI6Imh0dHBzOi8vd3d3LmQtdHJ1c3QubmV0L2RlL2FnYiIsIlFDX2hhc2giOiI3UXptNUVqdXpYS1NIRmxjME9IOVBQOXFVYUgtVkJsMmFHTmJ3WWoxb09BIiwiUUNfaGFzaEFsZ29yaXRobU9JRCI6IjIuMTYuODQwLjEuMTAxLjMuNC4yLjEiLCJ0cmFuc2FjdGlvbl9kYXRhX2hhc2hlc19hbGciOlsic2hhLTI1NiJdfQ"
-                        ]
-                    }
-                """.trimIndent()
             }
         }
     } - {
@@ -216,10 +107,9 @@ val KeyBindingTests by matrixSuite {
 
             verifierOid4Vp.validateAuthnResponse(authnResponse.url).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.map {
-                    it.getOrThrow()
-                }.shouldBeSingleton().first()
+                .shouldBeInstanceOf<VpTokenValidationResultDCQL>()
+                .credentialQueryResponseValidations.values.flatMap { it.map { it.getOrThrow() } }
+                .shouldBeSingleton().first()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                 .sdJwtSigned.keyBindingJws.shouldNotBeNull().payload.apply {
                     transactionDataHashes.shouldNotBeNull()
@@ -251,10 +141,9 @@ val KeyBindingTests by matrixSuite {
 
             val result = verifierOid4Vp.validateAuthnResponse(authnResponse.url).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.map {
-                    it.getOrThrow()
-                }.shouldBeSingleton().first()
+                .shouldBeInstanceOf<VpTokenValidationResultDCQL>()
+                .credentialQueryResponseValidations.values.flatMap { it.map { it.getOrThrow() } }
+                .shouldBeSingleton().first()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
 
             with(result.sdJwtSigned.keyBindingJws.shouldNotBeNull().payload) {
@@ -285,8 +174,8 @@ val KeyBindingTests by matrixSuite {
 
             verifierOid4Vp.validateAuthnResponse(malignResponse.params.formUrlEncode()).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.shouldBeSingleton().first().isFailure shouldBe true
+                .shouldBeInstanceOf<VpTokenValidationResultDCQL>()
+                .credentialQueryResponseValidations.values.shouldBeSingleton().first().first().isFailure shouldBe true
         }
 
         "Transaction Data validation can be turned off" {
@@ -315,51 +204,10 @@ val KeyBindingTests by matrixSuite {
 
             lenientVerifier.validateAuthnResponse(malignResponse.params.formUrlEncode()).getOrThrow()
                 .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.map {
-                    it.getOrThrow()
-                }.shouldBeSingleton().first()
+                .shouldBeInstanceOf<VpTokenValidationResultDCQL>()
+                .credentialQueryResponseValidations.values.flatMap { it.map { it.getOrThrow() } }
+                .shouldBeSingleton().first()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-        }
-
-        "Hash of transaction data is not changed during processing" {
-            val referenceHash = it.cibaWalletTransactionData.toByteArray(Charsets.UTF_8).sha256()
-            val validNonces = mutableSetOf<String>()
-
-            val verifierOid4Vp = OpenId4VpVerifier(
-                keyMaterial = EphemeralKeyWithoutCert(),
-                clientIdScheme = ClientIdScheme.RedirectUri(it.clientId),
-                stateToAuthnRequestStore = it.externalMapStore,
-                nonceService = object : NonceService {
-                    override suspend fun provideNonce() = uuid4().toString().also(validNonces::add)
-                    override suspend fun verifyNonce(it: String) = it in validNonces
-                    override suspend fun verifyAndRemoveNonce(it: String) = validNonces.remove(it)
-                },
-            )
-
-            val state = it.holderOid4vp.startAuthorizationResponsePreparation(it.cibaWalletTestVector)
-                .getOrThrow().apply {
-                    request.parameters.transactionData.shouldNotBeEmpty().shouldNotBeNull()
-                }
-
-            validNonces += state.request.parameters.nonce.shouldNotBeNull()
-            it.externalMapStore.put(
-                "iTGlKl-AJxmncWPbXHp2xy58bNy18wqZ4TR9EzhBl2R4ulxeTEO0VyWYR2qMDpCDV5JWeOxecTqcEJ61bFKrUg",
-                state.request.parameters,
-            )
-
-            val authnResponse = it.holderOid4vp.createAuthnResponse(state.request).getOrThrow()
-                .shouldBeInstanceOf<AuthenticationResponseResult.Post>()
-
-            verifierOid4Vp.validateAuthnResponse(authnResponse.params.formUrlEncode()).getOrThrow()
-                .vpTokenValidationResult.shouldNotBeNull().getOrThrow()
-                .shouldBeInstanceOf<VpTokenValidationResultPresentationExchange>()
-                .inputDescriptorResponseValidations.values.map {
-                    it.getOrThrow()
-                }.shouldBeSingleton().first()
-                .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-                .sdJwtSigned.keyBindingJws.shouldNotBeNull().payload.transactionDataHashes!!.first()
-                .shouldBe(referenceHash)
         }
     }
 }